computeMean() returns a double

+++ src/main/cpp/include/common/Accumulator.h
@@ -15,7 +15,7 @@ struct Accumulator {
 
     Accumulator(size_t capacity);
 
-    int64_t computeMean() const;
+    double computeMean() const;
     void push(int64_t val);
 };
